Stop exposing body provider implementations
